add ChiVinh payment order_detail

+++ lib/Views/CategoryManager.dart
@@ -37,14 +37,14 @@ class _CategoryMangerState extends State<CategoryManger> {
                 Navigator.pushReplacement(
                     context, MaterialPageRoute(builder: (context) => Index()));
               },
-              icon: Icon(Icons.arrow_back)),
+              icon: const Icon(Icons.arrow_back)),
         ),
         floatingActionButton: FloatingActionButton(
           onPressed: () {
             Navigator.pushReplacement(context,
-                MaterialPageRoute(builder: (context) => CategoryAdd()));
+                MaterialPageRoute(builder: (context) => const CategoryAdd()));
           },
-          child: Icon(Icons.add),
+          child: const Icon(Icons.add),
         ),
         floatingActionButtonLocation: FloatingActionButtonLocation.endTop,
         body: FutureBuilder(
@@ -75,13 +75,52 @@ class _CategoryMangerState extends State<CategoryManger> {
                                                         theloai: categories[
                                                             index])));
                                       },
-                                      icon: Icon(Icons.edit)),
+                                      icon: const Icon(Icons.edit)),
                                   IconButton(
-                                      onPressed: () async {
-                                        await deleteCategories(
-                                            categories[index]['id']);
+                                      onPressed: () {
+                                        bool result = false;
+                                        showDialog<void>(
+                                          context: context,
+                                          barrierDismissible:
+                                              false, // user must tap button!
+                                          builder: (BuildContext context) {
+                                            return AlertDialog(
+                                              title: const Text('Xác nhận xóa'),
+                                              content:
+                                                  const SingleChildScrollView(
+                                                child: ListBody(
+                                                  children: <Widget>[
+                                                    Text(
+                                                        'Bạn có chắc chắn muốn xóa thể loại này không?'),
+                                                  ],
+                                                ),
+                                              ),
+                                              actions: <Widget>[
+                                                TextButton(
+                                                  child: const Text('Hủy'),
+                                                  onPressed: () {
+                                                    Navigator.of(context).pop();
+                                                  },
+                                                ),
+                                                TextButton(
+                                                  child: const Text('Xóa'),
+                                                  onPressed: () {
+                                                    deleteCategories(
+                                                        categories[index]
+                                                            ['id']);
+                                                    result = true;
+                                                    Navigator.of(context).pop();
+                                                  },
+                                                ),
+                                              ],
+                                            );
+                                          },
+                                        );
+                                        if (result) {
+                                          setState(() {});
+                                        }
                                       },
-                                      icon: Icon(Icons.delete)),
+                                      icon: const Icon(Icons.delete)),
                                 ],
                               ),
                             ),
@@ -90,7 +129,7 @@ class _CategoryMangerState extends State<CategoryManger> {
                       );
                     },
                   )
-                : CircularProgressIndicator();
+                : const CircularProgressIndicator();
           },
         ));
   }
